refactor(features): flatten component folder structures
Remove redundant nested folders for workspace-manager, ui-manager, and
job-runner components, and update corresponding import statements
across the codebase.
