+++ src/features/ui-manager/components/menu-bar/UiInspectorButton.tsx
