+++ src/features/ui-manager/components/side-panels/UiControlsPanel.tsx
