+++ src/features/ui-manager/components/side-panels/UiInspectorPanel.tsx
