+++ src/features/ui-manager/components/tool-bar/UiInspectorButton.tsx
