+++ src/features/workspace-manager/components/menu-bar/WorkspaceCloseButton.tsx
