+++ src/features/workspace-manager/components/menu-bar/WorkspaceCopyButton.tsx
