+++ src/features/workspace-manager/components/menu-bar/WorkspaceNewButton.tsx
