+++ src/features/workspace-manager/components/menu-bar/WorkspaceOpenButton.tsx
