+++ src/features/workspace-manager/components/menu-bar/WorkspaceRenameButton.tsx
