+++ src/layouts/workbench-window/application-header/menu-bar/file-submenu/MenuBarFileSubmenu.tsx
@@ -1,13 +1,13 @@
-import WorkspaceNewButton from "@/features/workspace-manager/components/menu-bar/WorkspaceNewButton/WorkspaceNewButton";
-import WorkspaceSaveButton from "@/features/workspace-manager/components/menu-bar/WorkspaceSaveButton/WorkspaceSaveButton";
-import WorkspaceSaveAsButton from "@/features/workspace-manager/components/menu-bar/WorkspaceSaveAsButton/WorkspaceSaveAsButton";
-import WorkspaceRenameButton from "@/features/workspace-manager/components/menu-bar/WorkspaceRenameButton/WorkspaceRenameButton";
-import WorkspaceCopyButton from "@/features/workspace-manager/components/menu-bar/WorkspaceCopyButton/WorkspaceCopyButton";
-import WorkspaceImportButton from "@/features/workspace-manager/components/menu-bar/WorkspaceImportButton/WorkspaceImportButton";
-import WorkspaceExportButton from "@/features/workspace-manager/components/menu-bar/WorkspaceExportButton/WorkspaceExportButton";
-import WorkspaceCloseButton from "@/features/workspace-manager/components/menu-bar/WorkspaceCloseButton/WorkspaceCloseButton";
+import WorkspaceNewButton from "@/features/workspace-manager/components/menu-bar/WorkspaceNewButton";
+import WorkspaceSaveButton from "@/features/workspace-manager/components/menu-bar/WorkspaceSaveButton";
+import WorkspaceSaveAsButton from "@/features/workspace-manager/components/menu-bar/WorkspaceSaveAsButton";
+import WorkspaceRenameButton from "@/features/workspace-manager/components/menu-bar/WorkspaceRenameButton";
+import WorkspaceCopyButton from "@/features/workspace-manager/components/menu-bar/WorkspaceCopyButton";
+import WorkspaceImportButton from "@/features/workspace-manager/components/menu-bar/WorkspaceImportButton";
+import WorkspaceExportButton from "@/features/workspace-manager/components/menu-bar/WorkspaceExportButton";
+import WorkspaceCloseButton from "@/features/workspace-manager/components/menu-bar/WorkspaceCloseButton";
 import MenuSeparator from "@/components/menu-bar/MenuBarSeparator";
-import WorkspaceOpenButton from "@/features/workspace-manager/components/menu-bar/WorkspaceOpenButton/WorkspaceOpenButton";
+import WorkspaceOpenButton from "@/features/workspace-manager/components/menu-bar/WorkspaceOpenButton";
 
 export default function MenuBarFileSubmenu() {
   return (
