+++ src/layouts/workbench-window/controls-panel/ControlsPanel.tsx
@@ -1,4 +1,4 @@
-import UiControlsPanel from "@/features/ui-manager/components/side-panels/UiControlsPanel/UiControlsPanel";
+import UiControlsPanel from "@/features/ui-manager/components/side-panels/UiControlsPanel";
 import AlgoSection from "./algo-section/AlgoSection";
 import EnvSection from "@/layouts/workbench-window/controls-panel/env-section/EnvSection";
 import ObjectSection from "@/layouts/workbench-window/controls-panel/object-section/ObjectSection";
